test(skills): isolate description limit regression coverage
