+++ src/agents/tools/skill-workshop-tool.test.ts
@@ -2,7 +2,6 @@
 // applying generated skills to the workspace.
 import fs from "node:fs/promises";
 import path from "node:path";
-import { validateToolArguments } from "openclaw/plugin-sdk/llm";
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
 import { consumeRunSkillUsage, recordRunSkillUsage } from "../../skills/runtime/run-usage.js";
 import { writeWorkspaceSkills } from "../../skills/test-support/e2e-test-helpers.js";
@@ -208,60 +207,6 @@ describe("skill_workshop tool", () => {
     expect(schema).toContain("artifact_path");
     expect(tool.description).toContain(lazyDescription);
     expect(tool.description).toContain(SKILL_AUTHORING_STANDARDS_PROMPT);
-  });
-
-  it("lets the proposal service explain overlong descriptions", async () => {
-    const workspaceDir = await tempDirs.make("openclaw-skill-workshop-description-limit-");
-    const tool = createSkillWorkshopTool({ workspaceDir, env: testState.env });
-    const args = {
-      action: "create" as const,
-      name: "Long Description",
-      description: "x".repeat(161),
-      proposal_content: "# Long Description\n",
-    };
-    const call = {
-      type: "toolCall" as const,
-      id: "call-long-description",
-      name: "skill_workshop",
-      arguments: args,
-    };
-    expect(validateToolArguments(tool, call)).toEqual(args);
-
-    await expect(tool.execute(call.id, args)).rejects.toThrow(
-      "Skill proposal description is too large (161 bytes, max 160).",
-    );
-  });
-
-  it("lets the proposal service explain overlong collection descriptions", async () => {
-    const workspaceDir = await tempDirs.make("openclaw-skill-collection-description-limit-");
-    const collectionReconcile = {};
-    const tool = createSkillWorkshopTool({ workspaceDir, env: testState.env, collectionReconcile });
-    const args = {
-      action: "reconcile" as const,
-      collection: [
-        {
-          action: "write" as const,
-          name: "long-description",
-          description: "x".repeat(161),
-          content: "# Long Description\n",
-        },
-      ],
-    };
-    const call = {
-      type: "toolCall" as const,
-      id: "call-long-collection-description",
-      name: "skill_workshop",
-      arguments: args,
-    };
-    expect(validateToolArguments(tool, call)).toEqual(args);
-
-    await expect(tool.execute(call.id, args)).rejects.toThrow(
-      "Skill proposal description is too large (161 bytes, max 160).",
-    );
-    await expect(
-      fs.access(path.join(workspaceDir, "skills", "long-description", "SKILL.md")),
-    ).rejects.toThrow();
-    expect(collectionReconcile).not.toHaveProperty("result");
   });
 
   it("evaluates an exact pending draft and exposes the persisted result", async () => {
